Refactoring the code - Created separate folders for each product - Changes in package.json
