+++ package.json
@@ -5,22 +5,22 @@
   "homepage": "https://dev.kone.com",
   "main": "src/examples/elevator-call-2-demo.ts",
   "scripts": {
-    "start": "ts-node src/examples/elevator-call-2-demo.ts",
-    "start:elevator-cancel-call-demo": "ts-node src/examples/elevator-cancel-call-demo.ts",
-    "start:monitoring-siteapi-demo": "ts-node src/examples/monitoring-siteapi-demo.ts",
-    "start:ping-demo": "ts-node src/examples/ping-demo.ts",
-    "start:config-demo": "ts-node src/examples/config-demo.ts",
-    "start:actions-demo": "ts-node src/examples/actions-demo.ts",
-    "start:call-and-monitoring-demo": "ts-node src/examples/call-and-monitoring-demo.ts",
-    "start:websocket-sessions-basic": "ts-node src/examples/websocket-sessions-basic.ts",
-    "start:websocket-sessions-advanced": "ts-node src/examples/websocket-sessions-advanced.ts",
-    "start:operational-apis-demo": "ts-node src/examples/operational-apis-demo.ts",
-    "start:get-equipment-availability-demo": "ts-node src/examples/get-equipment-availability-demo.ts",
-    "start:get-equipment-status-demo": "ts-node src/examples/get-equipment-status.ts",
-    "start:get-elevator-movement-event-demo": "ts-node src/examples/get-elevator-movement-event.ts",
-    "start:get-elevator-door-button-event-demo": "ts-node src/examples/get-elevator-door-button-event.ts",
-    "start:service-info-2-restapi-demo": "ts-node src/examples/service-info-2-restapi-demo.ts",
-    "start:get-escalator-event-demo": "ts-node src/examples/get-escalator-event-demo.ts",
+    "start": "ts-node src/examples/elevator-call-api/elevator-call-2-demo.ts",
+    "start:elevator-cancel-call-demo": "ts-node src/examples/elevator-call-api/elevator-cancel-call-demo.ts",
+    "start:monitoring-siteapi-demo": "ts-node src/examples/elevator-call-api/monitoring-siteapi-demo.ts",
+    "start:ping-demo": "ts-node src/examples/elevator-call-api/ping-demo.ts",
+    "start:config-demo": "ts-node src/examples/elevator-call-api/config-demo.ts",
+    "start:actions-demo": "ts-node src/examples/elevator-call-api/actions-demo.ts",
+    "start:call-and-monitoring-demo": "ts-node src/examples/elevator-call-api/call-and-monitoring-demo.ts",
+    "start:websocket-sessions-basic": "ts-node src/examples/elevator-call-api/websocket-sessions-basic.ts",
+    "start:websocket-sessions-advanced": "ts-node src/examples/elevator-call-api/websocket-sessions-advanced.ts",
+    "start:operational-apis-demo": "ts-node src/examples/operational-apis/operational-apis-demo.ts",
+    "start:get-equipment-availability-demo": "ts-node src/examples/operational-apis/euipment-status-api/get-equipment-availability-demo.ts",
+    "start:get-equipment-status-demo": "ts-node src/examples/operational-apis/euipment-status-api/get-equipment-status.ts",
+    "start:get-elevator-movement-event-demo": "ts-node src/examples/operational-apis/euipment-status-api/get-elevator-movement-event.ts",
+    "start:get-elevator-door-button-event-demo": "ts-node src/examples/operational-apis/euipment-status-api/get-elevator-door-button-event.ts",
+    "start:service-info-2-restapi-demo": "ts-node src/examples/operational-apis/service-info-api/service-info-2-restapi-demo.ts",
+    "start:get-escalator-event-demo": "ts-node src/examples/operational-apis/euipment-status-api/get-escalator-event-demo.ts",
     "build": "npm run clean && tsc -b .",
     "clean": "rimraf dist"
   },
