+++ src/examples/elevator-call-api/config-demo.ts
@@ -7,7 +7,7 @@ import {
   fetchAccessToken,
   openWebSocketConnection,
   validateClientIdAndClientSecret,
-} from '../common/koneapi'
+} from '../../common/koneapi'
 
 /**
  * Update these two variables with your own credentials or set them up as environment variables.
@@ -15,6 +15,7 @@ import {
 const CLIENT_ID: string = process.env.CLIENT_ID || 'YOUR_CLIENT_ID' // eg. 'dcf48ab0-a902-4b52-8c53-1a9aede716e5'
 const CLIENT_SECRET: string = process.env.CLIENT_SECRET || 'YOUR_CLIENT_SECRET' // eg. '31d1329f8344fc12b1a960c8b8e0fc6a22ea7c35774c807a4fcabec4ffc8ae5b'
 const BUILDING_ID: string = process.env.BUILDING_ID || ''
+
 /**
  * Function is used to log out incoming WebSocket messages
  *
@@ -24,7 +25,9 @@ const onWebSocketMessage = (data: string): void => {
   let dataBlob = JSON.parse(data)
 
   console.log('Incoming WebSocket message', dataBlob)
-  console.log(dataBlob?.data?.call_types)
+  console.log(dataBlob?.data?.groups)
+  console.log(dataBlob?.data?.destinations)
+
   console.log('timing ' + new Date())
 }
 
@@ -39,6 +42,7 @@ const start = async () => {
   let accessToken = await fetchAccessToken(CLIENT_ID, CLIENT_SECRET, [
     'application/inventory',
     `callgiving/group:${BUILDING_ID}:1`,
+    // `callgiving/group:${BUILDING_ID}:2`,
   ])
   console.log('AccessToken successfully fetched')
 
@@ -56,10 +60,9 @@ const start = async () => {
   // Build the call payload using the areas previously generated
   const destinationCallPayload: any = {
     type: 'common-api',
-    buildingId: targetBuildingId,
-    callType: 'actions',
     requestId: uuidv4(),
-    // callType: 'config',
+    buildingId: targetBuildingId,
+    callType: 'config',
     groupId: '1',
   }
