+++ src/examples/elevator-call-api/elevator-call-2-demo.ts
@@ -7,17 +7,15 @@ import {
   fetchAccessToken,
   openWebSocketConnection,
   validateClientIdAndClientSecret,
-} from '../common/koneapi'
+} from '../../common/koneapi'
 
 /**
  * Update these two variables with your own credentials or set them up as environment variables.
  */
 const CLIENT_ID: string = process.env.CLIENT_ID || 'YOUR_CLIENT_ID' // eg. 'dcf48ab0-a902-4b52-8c53-1a9aede716e5'
 const CLIENT_SECRET: string = process.env.CLIENT_SECRET || 'YOUR_CLIENT_SECRET' // eg. '31d1329f8344fc12b1a960c8b8e0fc6a22ea7c35774c807a4fcabec4ffc8ae5b'
-const BUILDING_ID: string = process.env.BUILDING_ID || ''
+const BUILDING_ID: string = 'Qt8lIGiut3t'
 
-let webSocketConnection: any
-let targetBuildingId: string
 /**
  * Function is used to log out incoming WebSocket messages
  *
@@ -27,38 +25,9 @@ const onWebSocketMessage = (data: string): void => {
   let dataBlob = JSON.parse(data)
 
   console.log('Incoming WebSocket message', dataBlob)
-  if (dataBlob.data.session_id) {
-    console.log(`Making monitoring request ${dataBlob.data.session_id}`)
-    activateMonitoring(dataBlob.data.session_id)
-  }
   console.log('timing ' + new Date())
 }
 
-function activateMonitoring(session_id: number) {
-  // Build the call payload using the areas previously generated
-  const monitoringCallPayload: any = {
-    type: 'site-monitoring',
-    requestId: uuidv4(),
-    buildingId: targetBuildingId,
-    callType: 'monitor',
-    // callType: 'config',
-    groupId: '1',
-    payload: {
-      sub: 'KCESmartLiftClient2',
-      duration: 300,
-      subtopics: [
-        `call_state/${session_id}/being_allocated`,
-        `call_state/${session_id}/being_served`,
-        `call_state/${session_id}/served_soon`,
-        `call_state/${session_id}/served`
-      ],
-    },
-  }
-
-  console.log('monitoring request payload')
-  console.log(JSON.stringify(monitoringCallPayload))
-  webSocketConnection.send(JSON.stringify(monitoringCallPayload))
-}
 /**
  * Main function that starts the script execution
  */
@@ -74,16 +43,17 @@ const start = async () => {
   console.log('AccessToken successfully fetched')
 
   // Select the first available building
-  targetBuildingId = `building:${BUILDING_ID}`
+  const targetBuildingId = `building:${BUILDING_ID}`
   // Fetch the topology of the specific building
 
   // Open the WebSocket connection
-  webSocketConnection = await openWebSocketConnection(accessToken)
+  const webSocketConnection = await openWebSocketConnection(accessToken)
   console.log('WebSocket open ' + new Date())
 
   // Add handler for incoming messages
   webSocketConnection.on('message', (data: any) => onWebSocketMessage(data))
 
+  // Build the call payload using the areas previously generated
   const destinationCallPayload: any = {
     type: 'lift-call-api-v2',
     buildingId: targetBuildingId,
@@ -91,11 +61,14 @@ const start = async () => {
     groupId: '1',
     payload: {
       request_id: getRequestId(),
-      area: 3000,
+      area: 5000,
       time: '2020-10-10T07:17:33.298515Z',
-      terminal: 1,
-      // action 2 is for DCS call
-      call: { action: 2, destination: 5000 },
+      terminal: 2,
+      // terminal: 10011,
+      call: {
+        action: 3,
+        destination: 3000,
+      },
     },
   }
