+++ src/examples/elevator-call-api/elevator-cancel-call-demo.ts
@@ -7,7 +7,7 @@ import {
   fetchAccessToken,
   openWebSocketConnection,
   validateClientIdAndClientSecret,
-} from '../common/koneapi'
+} from '../../common/koneapi'
 
 /**
  * Update these two variables with your own credentials or set them up as environment variables.
@@ -16,6 +16,8 @@ const CLIENT_ID: string = process.env.CLIENT_ID || 'YOUR_CLIENT_ID' // eg. 'dcf4
 const CLIENT_SECRET: string = process.env.CLIENT_SECRET || 'YOUR_CLIENT_SECRET' // eg. '31d1329f8344fc12b1a960c8b8e0fc6a22ea7c35774c807a4fcabec4ffc8ae5b'
 const BUILDING_ID: string = process.env.BUILDING_ID || ''
 
+let webSocketConnection: any
+let targetBuildingId: string
 /**
  * Function is used to log out incoming WebSocket messages
  *
@@ -25,9 +27,27 @@ const onWebSocketMessage = (data: string): void => {
   let dataBlob = JSON.parse(data)
 
   console.log('Incoming WebSocket message', dataBlob)
+  if (dataBlob.data.session_id) {
+    console.log('Making delete call request')
+    makeDeleteCallRequest(dataBlob.data.session_id)
+  }
   console.log('timing ' + new Date())
 }
 
+function makeDeleteCallRequest(sessionId: any) {
+  const deleteCallPayload: any = {
+    type: 'lift-call-api-v2',
+    buildingId: targetBuildingId,
+    callType: 'delete',
+    groupId: '1',
+    payload: {
+      session_id: sessionId,
+    },
+  }
+  console.log('delete request payload')
+  console.log(JSON.stringify(deleteCallPayload))
+  webSocketConnection.send(JSON.stringify(deleteCallPayload))
+}
 /**
  * Main function that starts the script execution
  */
@@ -43,11 +63,11 @@ const start = async () => {
   console.log('AccessToken successfully fetched')
 
   // Select the first available building
-  const targetBuildingId = `building:${BUILDING_ID}`
+  targetBuildingId = `building:${BUILDING_ID}`
   // Fetch the topology of the specific building
 
   // Open the WebSocket connection
-  const webSocketConnection = await openWebSocketConnection(accessToken)
+  webSocketConnection = await openWebSocketConnection(accessToken)
   console.log('WebSocket open ' + new Date())
 
   // Add handler for incoming messages
@@ -55,21 +75,17 @@ const start = async () => {
 
   // Build the call payload using the areas previously generated
   const destinationCallPayload: any = {
-    type: 'site-monitoring',
-    requestId: uuidv4(),
+    type: 'lift-call-api-v2',
     buildingId: targetBuildingId,
-    callType: 'monitor',
-    // callType: 'config',
+    callType: 'action',
     groupId: '1',
     payload: {
-      sub: 'somethinaaa1',
-      duration: 300,
-      // subtopics: ['action/+/+'],
-      // subtopics: ['+/next_stop_eta'],
-      // subtopics: ['+/stopping'],
-      // subtopics: ['call_state/+/+'],
-      // subtopics: ['+/status'],
-      subtopics: ['+/position','+/doors'],
+      request_id: getRequestId(),
+      area: 3000,
+      time: '2020-10-10T07:17:33.298515Z',
+      terminal: 1,
+      // action 2 is for DCS call
+      call: { action: 2, destination: 5000 },
     },
   }
 
@@ -78,4 +94,7 @@ const start = async () => {
   webSocketConnection.send(JSON.stringify(destinationCallPayload))
 }
 
+function getRequestId() {
+  return Math.floor(Math.random() * 1000000000)
+}
 start()
