+++ src/examples/elevator-call-api/monitoring-siteapi-demo.ts
@@ -7,7 +7,7 @@ import {
   fetchAccessToken,
   openWebSocketConnection,
   validateClientIdAndClientSecret,
-} from '../common/koneapi'
+} from '../../common/koneapi'
 
 /**
  * Update these two variables with your own credentials or set them up as environment variables.
@@ -15,6 +15,7 @@ import {
 const CLIENT_ID: string = process.env.CLIENT_ID || 'YOUR_CLIENT_ID' // eg. 'dcf48ab0-a902-4b52-8c53-1a9aede716e5'
 const CLIENT_SECRET: string = process.env.CLIENT_SECRET || 'YOUR_CLIENT_SECRET' // eg. '31d1329f8344fc12b1a960c8b8e0fc6a22ea7c35774c807a4fcabec4ffc8ae5b'
 const BUILDING_ID: string = process.env.BUILDING_ID || ''
+
 /**
  * Function is used to log out incoming WebSocket messages
  *
@@ -35,7 +36,10 @@ const start = async () => {
 
   // Fetch the access token with both application/inventory scope and access to execute elevator calls on any building
   // accessible to the application - note that if you have many (100+) resources, you cannot use wildcards
-  let accessToken = await fetchAccessToken(CLIENT_ID, CLIENT_SECRET, ['application/inventory', 'callgiving/*'])
+  let accessToken = await fetchAccessToken(CLIENT_ID, CLIENT_SECRET, [
+    'application/inventory',
+    `callgiving/group:${BUILDING_ID}:1`,
+  ])
   console.log('AccessToken successfully fetched')
 
   // Select the first available building
@@ -51,14 +55,22 @@ const start = async () => {
 
   // Build the call payload using the areas previously generated
   const destinationCallPayload: any = {
-    type: 'common-api',
-    // requestId: uuidv4(),
+    type: 'site-monitoring',
+    requestId: uuidv4(),
     buildingId: targetBuildingId,
-    callType: 'ping',
-    payload: {
-      request_id: getRequestId(),
-    },
+    callType: 'monitor',
+    // callType: 'config',
     groupId: '1',
+    payload: {
+      sub: 'somethinaaa1',
+      duration: 300,
+      // subtopics: ['action/+/+'],
+      // subtopics: ['+/next_stop_eta'],
+      // subtopics: ['+/stopping'],
+      // subtopics: ['call_state/+/+'],
+      // subtopics: ['+/status'],
+      subtopics: ['+/position','+/doors'],
+    },
   }
 
   console.log(destinationCallPayload)
@@ -66,7 +78,4 @@ const start = async () => {
   webSocketConnection.send(JSON.stringify(destinationCallPayload))
 }
 
-function getRequestId() {
-  return Math.floor(Math.random() * 1000000000)
-}
 start()
