+++ src/examples/elevator-call-api/ping-demo.ts
@@ -7,15 +7,14 @@ import {
   fetchAccessToken,
   openWebSocketConnection,
   validateClientIdAndClientSecret,
-} from '../common/koneapi'
+} from '../../common/koneapi'
 
 /**
  * Update these two variables with your own credentials or set them up as environment variables.
  */
 const CLIENT_ID: string = process.env.CLIENT_ID || 'YOUR_CLIENT_ID' // eg. 'dcf48ab0-a902-4b52-8c53-1a9aede716e5'
 const CLIENT_SECRET: string = process.env.CLIENT_SECRET || 'YOUR_CLIENT_SECRET' // eg. '31d1329f8344fc12b1a960c8b8e0fc6a22ea7c35774c807a4fcabec4ffc8ae5b'
-const BUILDING_ID: string = 'Qt8lIGiut3t'
-
+const BUILDING_ID: string = process.env.BUILDING_ID || ''
 /**
  * Function is used to log out incoming WebSocket messages
  *
@@ -36,10 +35,7 @@ const start = async () => {
 
   // Fetch the access token with both application/inventory scope and access to execute elevator calls on any building
   // accessible to the application - note that if you have many (100+) resources, you cannot use wildcards
-  let accessToken = await fetchAccessToken(CLIENT_ID, CLIENT_SECRET, [
-    'application/inventory',
-    `callgiving/group:${BUILDING_ID}:1`,
-  ])
+  let accessToken = await fetchAccessToken(CLIENT_ID, CLIENT_SECRET, ['application/inventory', 'callgiving/*'])
   console.log('AccessToken successfully fetched')
 
   // Select the first available building
@@ -55,21 +51,14 @@ const start = async () => {
 
   // Build the call payload using the areas previously generated
   const destinationCallPayload: any = {
-    type: 'lift-call-api-v2',
+    type: 'common-api',
+    // requestId: uuidv4(),
     buildingId: targetBuildingId,
-    callType: 'action',
-    groupId: '1',
+    callType: 'ping',
     payload: {
       request_id: getRequestId(),
-      area: 5000,
-      time: '2020-10-10T07:17:33.298515Z',
-      terminal: 2,
-      // terminal: 10011,
-      call: {
-        action: 3,
-        destination: 3000,
-      },
     },
+    groupId: '1',
   }
 
   console.log(destinationCallPayload)
