+++ src/examples/operational-apis/service-info-api/service-info-2-restapi-demo.ts
@@ -2,8 +2,8 @@ import * as dotenv from 'dotenv'
 dotenv.config()
 import { v4 as uuidv4 } from 'uuid'
 import _ from 'lodash'
-import { fetchAccessToken, fetchResources, validateClientIdAndClientSecret } from '../common/koneapi'
-import { fetchServiceOrdersList, fetchSingleServiceOrder } from '../common/service-info-2-functions'
+import { fetchAccessToken, fetchResources, validateClientIdAndClientSecret } from '../../../common/koneapi'
+import { fetchServiceOrdersList, fetchSingleServiceOrder } from '../../../common/service-info-2-functions'
 
 /**
  * Update these two variables with your own credentials or set them up as environment variables.
